requirements.txt
mlflow==2.16.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
fastapi==0.115.5
uvicorn[standard]==0.32.0
prometheus-client==0.21.0
evidently==0.4.33
psycopg2-binary==2.9.10
PyYAML==6.0.2
dvc==3.56.0
black==24.10.0
ruff==0.7.1
pytest==8.3.3
nbconvert==7.16.4
pyarrow==17.0.0
joblib==1.4.2
